refactor(accordion-v4): update to new tokens
