--- packages/forma-36-react-components/src/components/Accordion/Accordion.css
@@ -1,73 +1,0 @@
-@import 'settings/colors';
-@import 'settings/typography';
-@import 'settings/transitions';
-@import 'settings/dimensions';
-
-.Accordion {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
-  list-style: none;
-
-  & .AccordionItem {
-    padding: 0;
-    margin: 0;
-    border-bottom: 1px solid var(--color-element-mid);
-
-    &:first-child {
-      border-top: 1px solid var(--color-element-mid);
-    }
-  }
-}
-
-.Accordion--start {
-  & .AccordionHeader svg {
-    min-width: 9px; /* necessary to align the chevron properly */
-    margin-right: var(--spacing-xs);
-  }
-}
-
-.Accordion--end {
-  & .AccordionHeader {
-    flex-direction: row-reverse;
-    justify-content: space-between;
-  }
-}
-
-.AccordionHeader {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  border: 0;
-  padding: var(--spacing-m);
-  background-color: transparent;
-  font-family: var(--font-stack-primary);
-  font-size: var(--font-size-l);
-  font-weight: var(--font-weight-demi-bold);
-  line-height: var(--line-height-default);
-  color: var(--color-text-base);
-  width: 100%;
-  cursor: pointer;
-  transition: background-color var(--transition-duration-default)
-    var(--transition-easing-default);
-
-  &:hover {
-    background-color: var(--color-element-lightest);
-  }
-  &:focus {
-    outline: none;
-    background-color: var(--color-element-lightest);
-  }
-}
-
-.AccordionHeader--expanded {
-  & .AccordionHeader__icon {
-    transform: rotate(180deg);
-  }
-}
-
-.AccordionHeader__icon {
-  transform: rotate(0deg);
-  transition: transform var(--transition-duration-default)
-    var(--transition-easing-default);
-}
