--- packages/forma-36-react-components/src/components/Accordion/AccordionPanel/AccordionPanel.css
@@ -1,16 +1,0 @@
-@import 'settings/dimensions';
-@import 'settings/transitions';
-
-.AccordionPanel {
-  box-sizing: border-box;
-  overflow: hidden;
-  height: 0;
-  transition: height var(--transition-duration-default)
-      var(--transition-easing-default),
-    padding var(--transition-duration-default) var(--transition-easing-default);
-}
-
-.AccordionPanel__content {
-  width: 100%;
-  padding: var(--spacing-xs) var(--spacing-m) var(--spacing-m);
-}
