--- packages/forma-36-react-components/src/components/Accordion/index.ts
@@ -1,8 +1,0 @@
-export { Accordion } from './Accordion';
-export type { AccordionProps } from './Accordion';
-export { AccordionHeader } from './AccordionHeader/AccordionHeader';
-export type { AccordionHeaderProps } from './AccordionHeader/AccordionHeader';
-export { AccordionItem } from './AccordionItem/AccordionItem';
-export type { AccordionItemProps } from './AccordionItem/AccordionItem';
-export { AccordionPanel } from './AccordionPanel/AccordionPanel';
-export type { AccordionPanelProps } from './AccordionPanel/AccordionPanel';
