+++ packages/forma-36-react-components/src/index.ts
@@ -93,9 +93,11 @@ export {
   SkeletonText,
   SkeletonRow,
 } from '@contentful/f36-skeleton';
-export * from './components/Accordion';
 export { Accordion, AccordionItem } from '@contentful/f36-accordion';
-// export * from './components/Accordion';
+export type {
+  AccordionProps,
+  AccordionItemProps,
+} from '@contentful/f36-accordion';
 export * from './components/Asset';
 export * from './components/Autocomplete';
 export * from './components/Card';
